+++ backend/Repositories/Travels/IExpenseRepository.cs
@@ -5,8 +5,8 @@ namespace backend.Repositories.Travels;
 public interface IExpenseRepository
 {
       Task<Expense> AddAsync(Expense expense);
-    Task<Expense?> GetByIdAsync(int expenseId);
-    Task<IReadOnlyCollection<Expense>> GetByAssigneeAsync(int employeeId);
-    Task<IReadOnlyCollection<Expense>> GetFilteredAsync(int? employeeId, int? travelId, DateTime? from, DateTime? to, string? status);
+    Task<Expense?> GetByIdAsync(long expenseId);
+    Task<IReadOnlyCollection<Expense>> GetByAssigneeAsync(long employeeId);
+    Task<IReadOnlyCollection<Expense>> GetFilteredAsync(long? employeeId, long? travelId, DateTime? from, DateTime? to, string? status);
     Task SaveAsync();
 }
